# docs/v2/customer/customer_swagger.yaml
definitions:
  customer.EntitiesResponseTemplate-customer_bankAccountResp:
    properties:
      count:
        type: integer
      next:
        type: string
      previous:
        type: string
      results:
        items:
          $ref: '#/definitions/customer.bankAccountResp'
        type: array
    type: object
...
  customer.EntitiesResponseTemplate-customer_guestBankAccountResp:
    properties:
      count:
        type: integer
      next:
        type: string
      previous:
        type: string
      results:
        items:
          $ref: '#/definitions/customer.guestBankAccountResp'
        type: array
    type: object
...
  customer.bankAccountResp:
    properties:
      account_number:
        type: string
      cash_in:
        type: number
      cash_out:
        type: number
      create_time:
        type: string
      customer_id:
        type: string
      id:
        type: string
      is_for_payment:
        type: boolean
      update_time:
        type: string
    type: object
  customer.bankAccountUpdateReq:
    properties:
      is_for_payment:
        type: boolean
    required:
    - is_for_payment
    type: object
...
  customer.errorResponse:
    properties:
      code:
        type: string
      detail:
        type: string
      message:
        type: string
    type: object
...
  customer.guestBankAccountResp:
    properties:
      account_number:
        type: string
      create_time:
        type: string
      customer_id:
        type: string
      id:
        type: string
      is_for_payment:
        type: boolean
      update_time:
        type: string
    type: object
...
  customer.meResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
...
  customer.tpBankbankAccountResp:
    properties:
      account_number:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/customer/v1/customers/{id}:
    get:
      consumes:
      - application/json
      description: Get a customer
      operationId: customer-get
      parameters:
      - description: ID of customer
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.meResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a customer
      tags:
      - Customer
...
  /api/customer/v1/me/bank-accounts:
    get:
      consumes:
      - application/json
      description: Show bank accounts
      operationId: bankaccount-listing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.EntitiesResponseTemplate-customer_bankAccountResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Show bank accounts
      tags:
      - Bank account
  /api/customer/v1/me/bank-accounts/guest:
    get:
      consumes:
      - application/json
      description: Show guest bank accounts
      operationId: guestbankaccount-listing
      parameters:
      - description: Bank account number
        in: query
        name: account_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.EntitiesResponseTemplate-customer_guestBankAccountResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Show guest bank accounts
      tags:
      - Bank account
...
  /me/bank-accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a bank account
      operationId: bankaccount-delete
      parameters:
      - description: ID of bank account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Delete a bank account
      tags:
      - Bank account
    get:
      consumes:
      - application/json
      description: Get a bank account
      operationId: bankaccount-get
      parameters:
      - description: ID of bank account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.bankAccountResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a bank account
      tags:
      - Bank account
    put:
      consumes:
      - application/json
      description: Update a bank account
      operationId: bankaccount-update
      parameters:
      - description: Update a bank account
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/customer.bankAccountUpdateReq'
      - description: ID of bank account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.bankAccountResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Update a bank account
      tags:
      - Bank account
  /me/bank-accounts/guest/{id}:
    get:
      consumes:
      - application/json
      description: Get a guest bank account
      operationId: guestbankaccount-get
      parameters:
      - description: ID of bank account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.guestBankAccountResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a guest bank account
      tags:
      - Bank account
  /me/bank-accounts/tp-bank/{account_number}:
    get:
      consumes:
      - application/json
      description: Get a tp bank bank account
      operationId: tpbankbankaccount-get
      parameters:
      - description: Bank account number
        in: query
        name: account_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.tpBankbankAccountResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a tp bank bank account
      tags:
      - Bank account
swagger: "2.0"